Fix Build.xml to delete test classes.
Sometimes when running tests on new patches  we were seeing a NoSuchMethodError
exception, leading the tests to fail.

In short this was because the ant clean task was not deleting the Test class
files in the test folder.  As such tests could be run with out of date Test
files, with may contain out of date refrences to the methods.

For example a patch may change the return type of Request.send() from void to
String.  As the test class files were not deleted when ant clean is run they
would not be rebuilt by ant test either, and as such would ty and call a send()
with return type void, which would throw an error as no such method would exist
in the new Request class.

FogBugz:

Bug:

Release-notes:

Manual:

Test-information:

1) Running ant clean now correctly deltes the class files in the test folder.

2) Applied Tarun Gupta's 'Adds Disco Features' patch to my existing stroke
directory which I had already built in.  This includes a change to
Request.send() return type from void to String.  After applying the patch I ran
ant clean and then ran ant test. The GetPrivateStorageRequestTest failed with a
NoSuchMethodError whenever the request.send() method was called.  After making
the changes to build.xml I ran ant clean and ant test again and the tests now
passed.

Change-Id: I577b587b6555c6c2f122b828bdf8d585774e862d
Reviewer: