Update findbugs version, and fix Makefile to fetch it
Version 3.0.1 of findbugs works with java 8 (the current version being used
breaks if you run tests under java 8). This change updates the Makefile so
that it fetches 3.0.1.

This change also updates the findbugs rules so that the findbugs
version is taken into account when checking to see if a version of
findbugs needs to be fetched. So this means that when this patch is
applied, a "make test" in existing checkouts will pull in the new
version of findbugs.

The same mechanism could be used for other 3rd party libraries when they get
changed.

Test-information:

Tested that new version of findbugs is fetched on a clean checkout, and for a
checkout that already has the previous findbugs.  Also tested that it won't
re-fetch findbugs if you've already got the right version.

Ran unit test under java 8, works OK
Ran unit test under java 7, works OK

Change-Id: I3515e62f6b663ec3d56619ab18d2bb262bb6cd25