Improve efficiency of JavaConnection reads.
Given that doRead() ends up populating a SafeByteArray as part of the
result is is more efficient to pass an empty SafeByteArray to doRead()
and push each chunk of data read directly to the it rather than use an
intermediate ByteArrayOutputStream. SafeByteArray.getData() will do any
concatenation necessary.

It is unnecessarily inefficient to use a class instance to return the
connection status from doRead. Remove JavaConnection.ReadResult.

Change-Id: I6c8dc000d9030d3929c71444eb5b489df93f1987