Changes to FileTransfer to be inline with current Swiften code
There has been some changes to the file transfer clases in Swiften since Tarun's
patch was written.  This patch updates the java classes to bring them in line
with the swiften one.

In details

IBBSendSession - added code to clean up onResponse connection when not needed.
In Swiften this was done by keeping reference to current session using boost
bindings to remove binding.  In java we have to keep a reference to the current
onResponse connection and disconnect it when not neeeded.

IncomingFileTransferManager - Remove redundant IQRouter field.

LocalJingeTransportCandidateGenerator -  Do not advertise link-local IPv6
addresses in FT candidates

OutgoingSIFileTransfer - File deleted

SOCKS5BytestreamProxiesManager and SOCKS5BytestreamProxyFinder - Search for all
proxies instead of just one.

Test-information:

Ran make and make test to check everything built ok and that the unit tests
still passed.

Change-Id: I696444e5074fe20625243693a44c836306b3a41e