Update FullPayloadSerializerCollection
Updates the FullPayloadSerializerCollection and
FullPayloadParserSerializerCollection to bring it inline with the Swiften
classes.  This included reordering the elements so they are in the same order as
the Swiften class (so its easier to see classes that are missing), and adding
ones that were missing.  For ones where no Java implemenation exits yet the line
is added but commented out.

Test-information:

Sanity test, ran unit tests all test still pass.

Change-Id: I318c8da577f638f4d799fabdfc09a3afa7fcb608