Add Block Parser and Serializer
Add BlockParser and BlockSerializer for the block payload classes.

This required slightly different approach to the Swiften code as C++ templates
are different to Java generics.  In short needed to add a type hierachy to the
Block payload classes and to the Parsers and Serilaizers associated with them.

Test-information:

Ran unit tests everything passed ok.

Change-Id: I3ea2d66afd7cb3b5c4c3515be3a1ef99d7dbb566