Add CombinedAvatarProviderTest
Adds the CombinedAvatarProviderTest class.

When I was adding the test I noticed that
testProviderUpdateWithAvatarDisappearingTriggersChange kept failing
due to a bug in the CombinedAvatarProvider class and
testRemoveProviderDisconnectsUpdates could not be implemented as the
class was incomplete.

I had to make some changes to CombinedAvatarProvider to fix these
issues.

Test-information:

Ran unit tests they all pass.

Change-Id: I0bfb68dd2b15df0f220f36c136aceadaf6545893