Finish porting S5BTransportSession
Finish porting S5BTransportSession to Stroke (previously a lot of it had been commented out).  To
do so had to introduce a heirachy to the SOCKS5BytestreamSession classes (which were being used as
type parameters for the Transport Session) and refactor some of the onByteSent signals to insure
they were all of the same type (Integer).

Also update PortingProgress.txt to update to give status of some of the porting files.

Test-information:

Ran unit tests they all still pass.

Change-Id: I4295b3a8829c208e65f5a46d19c35090f8c55865