Remove duplicate classes.
There ware duplicate copies of DummyFileTransferManager and
DummyFileTransferTransporterFactory in both stroke src and stroke test
directories (but within the same package).  This was confusing and
could cause problems when we come to update the files.

This patch removes the copies in the test directory, leaving only the
the src directory ones.

Test-information: Ran 'make test' everything still builds ok and all
tests pass.

Change-Id: I112d2b21e1217dcacd619393f81ba008a097e83a