Try to fix possible race condition in Connector and ProxiedConnection.
As per patch 'Fix possible race condition between Connection and
Connectors' (7eec2000d72f8fa597398704121d0b73a84ca284).

The issue occurs with ProxiedConnection that started connecting but do
not have an external reference anymore.  As soon as the handlers of
the ProxiedConnection are disconnected from the signals of the
connection_ object, the remaining references to a shared
ProxiedConnection vanish and the ProxiedConnection is deleted, while
it still requires access to its members in
ProxiedConnection::handleConnectFinished().

Test-information: Unit tests pass ok.

Change-Id: I6c81009262dd51cda17b1b93a15edf968f40e464