Rework Signals for time and space optimizations
Each Signal* class extends BaseSignal. Each Slot* class extends
BaseSlot.

BaseSignal manages the set of binds associated with it via addBind() and
getBinds() which are called from Signal*. It is optimized for the cases
of zero or one bind, only allocating a HashMap to hold a larger set when
needed.

The interaction with SignalConnection to handle disconnection is
effected via a callback interface rather than another Signal.

Change-Id: Ifa44c1eb40b778c303db947a6e74fe20d1b41a90